+++ Tests/PackageLoadingTests/PDNextVersionLoadingTests.swift
@@ -83,4 +83,23 @@ class PackageDescriptionNextVersionLoadingTests: PackageDescriptionLoadingTests 
             XCTAssertEqual(manifest.targets[0].pluginCapability, .postbuild)
         }
     }
+    
+    func testPackageDependencies() throws {
+        let stream = BufferedOutputByteStream()
+        stream <<< """
+            import PackageDescription
+            let package = Package(
+               name: "Foo",
+               dependencies: [
+                   .package(url: "/foo5", branch: "main"),
+                   .package(url: "/foo7", revision: "58e9de4e7b79e67c72a46e164158e3542e570ab6"),
+               ]
+            )
+            """
+        loadManifest(stream.bytes, toolsVersion: ToolsVersion(string: "999.0")) { manifest in
+        let deps = Dictionary(uniqueKeysWithValues: manifest.dependencies.map{ ($0.identity.description, $0) })
+            XCTAssertEqual(deps["foo5"], .scm(location: "/foo5", requirement: .branch("main")))
+            XCTAssertEqual(deps["foo7"], .scm(location: "/foo7", requirement: .revision("58e9de4e7b79e67c72a46e164158e3542e570ab6")))
+        }
+    }
 }